backend/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.7.0
attrs==24.2.0
bitarray==3.0.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
ckzg==2.0.1
click==8.1.7
cytoolz==1.0.0
eth-account==0.13.4
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.6.0
eth-rlp==2.1.0
eth-typing==5.0.1
eth-utils==5.1.0
eth_abi==5.1.0
fastapi==0.115.6
frozenlist==1.5.0
google-ai-generativelanguage==0.6.10
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
grpcio==1.68.1
grpcio-status==1.68.1
h11==0.14.0
h2==4.1.0
hexbytes==1.2.1
hpack==4.0.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.25.2
hyperframe==6.0.1
idna==3.10
iniconfig==2.0.0
multidict==6.1.0
numpy==2.1.3
packaging==24.2
pandas==2.2.3
parsimonious==0.10.0
pipe==2.2
pluggy==1.5.0
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycryptodome==3.21.0
pydantic==2.10.3
pydantic_core==2.27.1
pyparsing==3.2.0
pytest==8.3.4
pytest-asyncio==0.24.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyunormalize==16.0.0
regex==2024.11.6
requests==2.32.3
rlp==4.0.1
rsa==4.9
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
subgrounds==1.9.1
toolz==1.0.0
tqdm==4.67.1
types-requests==2.32.0.20241016
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.32.1
web3==7.6.0
websockets==13.1
yarl==1.18.3
